pyproject.toml
[project]
name = "tripweaver"
version = "0.1.0"
description = "Extensible travel itinerary planning backend."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "volcengine-python-sdk[ark]",
    "pydantic-settings>=2.10.0",
    "structlog>=25.4.0",
    "uvicorn[standard]>=0.35.0",
    "tavily>=1.1.0",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.23",
    "redis>=5.0.1",
    "pyjwt>=2.8.0",
    "shortuuid>=1.0.11",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tripweaver"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
